requirements.txt
requests
sparql-client
-e git+git://github.com/blazegraph/blazegraph-python.git#egg=pymantic
